backend/requirements.txt
# django related
django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.9.0
drf-spectacular==0.19.0

# images
Pillow==9.0.1

# local enviroment
python-dotenv==0.19.0

# wsgi server
gunicorn==20.1.0

# heroku
psycopg2-binary==2.9.1
dj-database-url==0.5.0

# ai
tensorflow==2.8.0

# others
bcrypt==3.2.0
python-dateutil==2.8.2
